hello/requirements.txt
python-dotenv
openai
colorama==0.4.6